pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "meta-growth-agent"
version = "0.1.0"
description = "Backend agent for Meta Ads diagnostics and competitor intelligence."
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sqlmodel>=0.0.22",
    "pydantic-settings>=2.4.0",
    "httpx>=0.27.0",
    "celery>=5.4.0",
    "redis>=5.0.0",
    "psycopg[binary]>=3.2.0",
    "structlog>=24.1.0",
    "python-dotenv>=1.0.1",
    "jinja2>=3.1.4",
    "anthropic>=0.34.0",
    "google-generativeai>=0.7.0",
    "pandas>=2.2.0",
    "orjson>=3.10.0",
    "weasyprint>=61.0",
    "tenacity>=9.0.0",
    "pendulum>=3.0.0",
    "aiofiles>=23.2.0",
    "opentelemetry-sdk>=1.25.0",
    "opentelemetry-instrumentation-fastapi>=0.46b0",
    "pyjwt>=2.9.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
    "ruff>=0.6.0",
    "mypy>=1.11.0",
    "types-redis",
    "types-requests",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
asyncio_mode = "auto"


